\name{fixef.gls}
\alias{fixef.gls}
\title{fixef gls}
\description{Fixed effects, i.e. coefficients of gls model}
\usage{
fixef.gls(object, ...)
}
\arguments{
  \item{object}{the fitted model}
  \item{\dots}{
}
}

\value{named numeric vector}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\code{\link{twNlme-package}}}

