\name{ranef.gls}
\alias{ranef.gls}
\title{ranef gls}
\description{Random effects, i.e. none, of gls model}
\usage{
ranef.gls(object, ...)
}
\arguments{
  \item{object}{the fitted model}
  \item{\dots}{
}
}

\value{numeric(0)}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\code{\link{twNlme-package}}}

