\name{seFixef.gls}
\alias{seFixef.gls}
\title{seFixef gls}
\description{Standard error of coefficients }
\usage{
seFixef.gls(object, ...)
}
\arguments{
  \item{object}{the fitted model}
  \item{\dots}{
}
}

\value{named numeric vector}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\code{\link{twNlme-package}}}

