\name{seFixef.lme}
\alias{seFixef.lme}
\title{seFixef lme}
\description{Standard error of fixed effects }
\usage{
seFixef.lme(object, ...)
}
\arguments{
  \item{object}{the fitted model}
  \item{\dots}{
}
}

\value{named numeric vector}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\code{\link{twNlme-package}}}

