\name{twNlme-internal}
\title{Internal twNlme objects}
\alias{.asFormulaTermVec}
\alias{.constructLinFormulaString}
\alias{.extractFixedComponent}
\alias{.extractFixedList}
\alias{.extractRandomList}
\alias{.fullNlmeCoefFormulas}
\alias{.getCategorialLevels}
\alias{.gsubFormTerm}
\alias{.tmp.f}
\description{Internal twNlme objects.}
\details{These are not to be called by the user.}
\keyword{internal}