\name{varFixef.gls}
\alias{varFixef.gls}
\title{varFixef gls}
\description{Variance-Covariance of gls effects}
\usage{\method{varFixef}{gls}(object, ...)}
\arguments{
  \item{object}{the fitted model}
  \item{\dots}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{twNlme-package}}}
\examples{FALSE}
