\name{varFixef.lme}
\alias{varFixef.lme}
\title{varFixef lme}
\description{Variance-Covariance of lme fixed effects}
\usage{\method{varFixef}{lme}(object, ...)}
\arguments{
  \item{object}{the fitted model}
  \item{\dots}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{twNlme-package}}}
\examples{FALSE}
