\name{varRanef.gls}
\alias{varRanef.gls}
\title{varRanef gls}
\description{Empty Variance-Covariance matrix for random effects}
\usage{
varRanef.gls(object, varName = NULL, ...)
}
\arguments{
  \item{object}{the fitted gls or gnls model}
  \item{varName}{for which parameter}
  \item{\dots}{
}
}

\value{0 object 0 matrix}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\code{\link{twNlme-package}}}

