\name{fixef.gls}
\alias{fixef.gls}
\title{fixef gls}
\description{Fixed effects, i.e. coefficients of gls model}
\usage{\method{fixef}{gls}(object, ...)}
\arguments{
  \item{object}{the fitted model}
  \item{\dots}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{twNlme-package}}}
\examples{FALSE}
