\name{ranef.gls}
\alias{ranef.gls}
\title{ranef gls}
\description{Random effects, i.e. none, of gls model}
\usage{\method{ranef}{gls}(object, ...)}
\arguments{
  \item{object}{the fitted model}
  \item{\dots}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{twNlme-package}}}
\examples{FALSE}
