\name{seFixef.gls}
\alias{seFixef.gls}
\title{seFixef gls}
\description{Standard error of coefficients }
\usage{\method{seFixef}{gls}(object, ...)}
\arguments{
  \item{object}{the fitted model}
  \item{\dots}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{twNlme-package}}}
\examples{FALSE}
