\name{seFixef.lme}
\alias{seFixef.lme}
\title{seFixef lme}
\description{Standard error of fixed effects }
\usage{\method{seFixef}{lme}(object, ...)}
\arguments{
  \item{object}{the fitted model}
  \item{\dots}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{twNlme-package}}}
\examples{FALSE}
