\name{varFixef.gls}
\alias{varFixef.gls}
\title{varFixef gls}
\description{Variance-Covariance of gls effects}
\usage{
varFixef.gls(object, ...)
}
\arguments{
  \item{object}{the fitted model}
  \item{\dots}{
}
}

\value{named numeric matrix}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\code{\link{twNlme-package}}}

