\name{varFixef.lme}
\alias{varFixef.lme}
\title{varFixef lme}
\description{Variance-Covariance of lme fixed effects}
\usage{
varFixef.lme(object, ...)
}
\arguments{
  \item{object}{the fitted model}
  \item{\dots}{
}
}

\value{named numeric matrix}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\code{\link{twNlme-package}}}

