\name{varRanef.gls}
\alias{varRanef.gls}
\title{varRanef gls}
\description{Empty Variance-Covariance matrix for random effects}
\usage{\method{varRanef}{gls}(object, varName = NULL, ...)}
\arguments{
  \item{object}{the fitted gls or gnls model}
  \item{varName}{for which parameter}
  \item{\dots}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{twNlme-package}}}
\examples{FALSE}
